wlan: Enable DXE FW logging channels if wlanLoggingEnable is set
When wlanLoggingEnable is not set then FW to host logging is disabled.
At present for enabling DXE RX FW logging channels wlanLoggingEnable ini
value is not validated. As a result FW logging channels are enabled
which are of no use if wlanLoggingEnable is disabled. So enable DXE
FW logging channels only if wlanLoggingEnable is set.

Change-Id: Ia7fdb96275c3627d2e523bbe8287e607e80be591
CRs-Fixed: 2418701